pyproject.toml
[project]
name = "ssm-mcp-server"
version = "0.1.0"
description = "MCP server for Cloudera Streams Messaging Manager (SMM)"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
authors = [
  { name = "SSM MCP", email = "ssm-mcp@example.com" }
]
dependencies = [
  "requests>=2.32.3",
  "pydantic>=2.8.2",
  "tenacity>=9.0.0",
  "anyio>=4.4.0",
  "mcp>=1.1.2",
  "fastmcp>=2.12.4",
  "pillow>=11.3.0",
  "python-dotenv>=1.0.0",
]

[project.scripts]
run-server = "ssm_mcp_server.server:main"

[project.optional-dependencies]
mcp = []

[tool.mcp]
servers = { ssm-mcp-server = "ssm_mcp_server.server:main" }

[tool.uv]
dev-dependencies = [
  "pytest>=8.2.0",
  "pytest-httpserver>=1.0.10",
  "ruff>=0.6.0",
]

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
